move files, data and update glue path
